Add content delete action and buttons to view

+++ app/helpers/application_helper.rb
@@ -1,20 +1,22 @@
 module ApplicationHelper
   def show_content_actions(page)
     if signed_in?
-      button_types(page).join(' ').html_safe
+      button_types(page)
     end
   end
 
   private
 
   def button_types(page)
-    if Content.find_by(page: page)
+    if page.nil?
       (link_to 'new', new_content_path, class: 'waves-effect waves-light btn-small')
-    else
+    elsif page
       [
         (link_to 'edit', edit_content_path(page), class: 'waves-effect waves-light btn-small'),
-        (link_to 'delete', content_path(page), method: :delete, class: 'waves-effect waves-light btn-small')
-      ]
+        (link_to 'delete', content_path(page), method: :delete, class: 'waves-effect waves-light btn-small', data: { confirm: 'Bist du sicher, dass du den Text löschen möchtest?'})
+      ].join(' ').html_safe
+    else
+      []
     end
   end
 end
